requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
python-dotenv==1.1.1
numpy==2.0.2
scikit-learn==1.5.2
transformers==4.56.1
datasets==4.0.0
peft==0.17.1
accelerate==1.10.1
torch>=2.4.0
bitsandbytes==0.47.0 ; platform_system != "Darwin"
tqdm==4.67.1
orjson==3.11.3
pytest==8.4.1
httpx==0.28.1
